requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pinecone-client>=5.0.0
PyPDF2==3.0.1
python-dotenv==1.0.0
pydantic==2.5.0

